python/requirements.txt
 
pillow==10.0.0
clip-interrogator==0.6.0
torch==2.0.1
torchvision==0.15.2